Interfaced P3 v3.1.6.6 into CM1 r21.1
This is the version that used operationally in the HRDPS (since IC-4).

The routines 'param.F' and 'mp_driver.F' were modified appropriately
to allow this P3 version to run using namelist option ptype = 51.

Currently, this also loads and advects the variables for zit_1 and qil_1;
but this is benign -- it just advects zero fields. It would be straigtfoward
to add options for the multicategory configuration of this P3 version.

Tested in CM1 (quarter_ss case) - it appears to work correctly.